requirements.txt
numpy
pillow
scikit-learn
xgboost